feat: create message-text component
