+++ projects/stream-chat-angular/src/lib/message/message.component.html
@@ -213,44 +213,26 @@
                   ) | translate
                 }}
               </div>
-              <div #messageTextElement data-testid="text">
-                <p>
-                  <ng-container *ngIf="messageTextParts; else defaultContent">
-                    <!-- eslint-disable-next-line @angular-eslint/template/use-track-by-function -->
-                    <ng-container *ngFor="let part of messageTextParts">
-                      <span
-                        *ngIf="part.type === 'text'; else mention"
-                        [innerHTML]="part.content"
-                      ></span>
-                      <ng-template #mention>
-                        <ng-template #defaultMention let-content="content">
-                          <span class="str-chat__message-mention">{{
-                            content
-                          }}</span>
-                        </ng-template>
-                        <ng-container
-                          *ngTemplateOutlet="
-                            (customTemplatesService.mentionTemplate$ | async) ||
-                              defaultMention;
-                            context: getMentionContext(part)
-                          "
-                        ></ng-container>
-                      </ng-template>
-                    </ng-container>
-                  </ng-container>
-                  <ng-template #defaultContent>
-                    <ng-container *ngIf="displayAs === 'text'; else asHTML">
-                      {{ messageText || "" }}
-                    </ng-container>
-                    <ng-template #asHTML
-                      ><span
-                        data-testid="html-content"
-                        [innerHTML]="messageText"
-                      ></span
-                    ></ng-template>
-                  </ng-template>
-                </p>
-              </div>
+              <ng-container
+                *ngTemplateOutlet="
+                  (customTemplatesService.messageTextTemplate$ | async) ||
+                    defaultText;
+                  context: getMessageTextContext()
+                "
+              ></ng-container>
+              <ng-template
+                #defaultText
+                let-message="message"
+                let-isQuoted="isQuoted"
+                let-shouldTranslate="shouldTranslate"
+              >
+                <stream-message-text
+                  [message]="message"
+                  [isQuoted]="isQuoted"
+                  [shouldTranslate]="shouldTranslate"
+                  data-testid="quoted-message-text"
+                ></stream-message-text>
+              </ng-template>
             </div>
           </div>
           <stream-icon-placeholder
@@ -357,14 +339,28 @@
           "
         ></ng-container>
       </ng-container>
-      <div
-        data-testid="quoted-message-text"
-        [innerHTML]="
-          message?.quoted_message?.translation ||
-          message?.quoted_message?.html ||
-          message?.quoted_message?.text
-        "
-      ></div>
+      <ng-container *ngIf="message?.quoted_message">
+        <ng-container
+          *ngTemplateOutlet="
+            (customTemplatesService.messageTextTemplate$ | async) ||
+              defaultText;
+            context: getQuotedMessageTextContext()
+          "
+        ></ng-container>
+        <ng-template
+          #defaultText
+          let-message="message"
+          let-isQuoted="isQuoted"
+          let-shouldTranslate="shouldTranslate"
+        >
+          <stream-message-text
+            [message]="message"
+            [isQuoted]="isQuoted"
+            [shouldTranslate]="shouldTranslate"
+            data-testid="quoted-message-text"
+          ></stream-message-text>
+        </ng-template>
+      </ng-container>
     </div>
   </div>
 </ng-template>
